more minor changes as self review

+++ software/controller/lib/hal/dma.cpp
@@ -185,7 +185,7 @@ void ChannelControl::Disable() {
   channel_reg->config.enable = 0;
 }
 
-uint32_t ChannelControl::Remaining() {
+size_t ChannelControl::Remaining() {
   auto *channel_reg = get_channel_reg(base_, channel_);
   return channel_reg->count;
 }
